Add gutter to top and right of heatmap
Prior to this change, a negative margin was used to offset the border so
that the main area aligns with the tick marks. The space added to the
height of the box due to the negative margin appears to be NOT accounted
for when distributing space to grid cells.

This change removes the negative margins and offsets the ticks mark with
an additional grid area that matches the size of the border. An element
is also required to occupy this area in order for tick mark boxes to be
automattically assigned to the correct cells.